refactor: moved js files out of js folder
