+++ frontend/assets/firebaseUtils.ts
